+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Changed
 - Rewrite run configuration
 - Tye global tool version is updated to 0.11.0-alpha.22111.1
+- Update to .NET 6
 
 ## [0.2.1]
 ### Added