+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 # tye-plugin Changelog
 
 ## [Unreleased]
+
+## [0.3.0]
 ### Added
 - New action to uninstall tye global tool
 - Ingress nodes